Refactor Project
- Refactor / add Project Settings
- Make Project compatible over eclipse and IntelliJ
