+++ .settings/org.eclipse.ltk.core.refactoring.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+org.eclipse.ltk.core.refactoring.enable.project.refactoring.history=false
